feat: parallel routes を用いて /home を再実装した
- Add ArticleCard component for displaying individual articles with hover effects.
- Create ArticleList component to render a list of articles or a message when no articles are found.
- Implement fetchArticles and fetchTagSearchHistoryArticles functions to retrieve articles from the API.
- Add styles for ArticleCard and ArticleList components.
- Develop ArticlesPage to fetch and display articles based on user preferences.
- Introduce ChatSidebar and ChatSideBarWrapper components for chat functionality.
- Implement responsive design for chat sidebar and toggle button.
- Create LogoutSection for user logout functionality with confirmation dialog.
- Refactor layout to accommodate new components and improve user experience.
- Remove obsolete HomeClient component and related styles.
