+++ culture-web/src/app/(authorized)/home/@articles/_drivers/fetchTagSearchHistoryArticles.ts
@@ -1,5 +1,5 @@
 import { apiClient } from '@/lib/apiClient'
-import type { Article } from '../../types'
+import type { Article } from '../../../types'
 
 export async function fetchTagSearchHistoryArticles(): Promise<Article[]> {
   try {
