+++ culture-web/src/app/(authorized)/home/@articles/_styles/articleCard.styles.ts
