+++ culture-web/src/app/(authorized)/home/@chatSideBar/_styles/chatSidebar.styles.ts
