+++ culture-web/src/app/(authorized)/home/_actions/signOutUser.ts
